add logging test and test fixtures
